+++ src/main/java/org/betonquest/betonquest/CustomDropListener.java
@@ -6,12 +6,11 @@ import org.betonquest.betonquest.exceptions.ObjectNotFoundException;
 import org.betonquest.betonquest.id.ItemID;
 import org.betonquest.betonquest.item.QuestItem;
 import org.bukkit.Bukkit;
+import org.bukkit.NamespacedKey;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.event.entity.EntityDeathEvent;
-import org.bukkit.metadata.MetadataValue;
-
-import java.util.List;
+import org.bukkit.persistence.PersistentDataType;
 
 @SuppressWarnings("PMD.CommentRequired")
 public class CustomDropListener implements Listener {
@@ -28,20 +27,20 @@ public class CustomDropListener implements Listener {
     @EventHandler(ignoreCancelled = true)
     public void onEntityDeath(final EntityDeathEvent event) {
         int dropIndex = 0;
-        while (event.getEntity().hasMetadata("betonquest-drops-" + dropIndex)) {
-            try {
-                final List<MetadataValue> metadata = event.getEntity().getMetadata("betonquest-drops-" + dropIndex);
-                for (final MetadataValue m : metadata) {
-                    final String item = m.asString();
-                    final int separatorIndex = item.indexOf(':');
-                    event.getDrops().add(new QuestItem(new ItemID(null, item.substring(0, separatorIndex)))
-                            .generate(Integer.parseInt(item.substring(separatorIndex + 1))));
+        String dataContainerValue;
+        do {
+            final NamespacedKey key = new NamespacedKey(BetonQuest.getInstance(), "betonquest-drops-" + dropIndex);
+            dataContainerValue = event.getEntity().getPersistentDataContainer().get(key, PersistentDataType.STRING);
+            if (dataContainerValue != null) {
+                final int separatorIndex = dataContainerValue.indexOf(':');
+                try {
+                    event.getDrops().add(new QuestItem(new ItemID(null, dataContainerValue.substring(0, separatorIndex)))
+                            .generate(Integer.parseInt(dataContainerValue.substring(separatorIndex + 1))));
+                } catch (final InstructionParseException | ObjectNotFoundException e) {
+                    log.warn("Error when dropping custom item from entity: " + e.getMessage(), e);
                 }
-            } catch (final InstructionParseException | ObjectNotFoundException e) {
-                log.warn("Error when dropping custom item from entity: " + e.getMessage(), e);
             }
             dropIndex++;
-        }
+        } while (dataContainerValue != null);
     }
-
 }