+++ CHANGELOG.md
@@ -21,6 +21,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   - `%player.display%` will use the display name used in chat
   - `%player.uuid%` will display the UUID of the player
 - marking of mobs now uses variables in all contexts of spawning killing and removing
+- marking of mobs is now stored in the persistent data container of the entity to prevent loss on server restart
 - default `default_conversation_IO` in config from `menu,chest` to `menu,tellraw`
 - corrected Russian messages in `messages.yml` file
 ### Deprecated